news/serializers.py
from rest_framework import serializers
from .models import News, Job


class NewsSerializer(serializers.ModelSerializer):
    class Meta:
        model = News
        fields = ('id', 'title', 'link', 'image', 'date_posted')


class JobSerializer(serializers.ModelSerializer):
    class Meta:
        model = Job
        fields = ('id', 'title', 'link')

